Moved test_tag_svg into test_tags
